+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Metadata.hs
@@ -1,7 +1,0 @@
-module Cardano.Ledger.Shelley.Metadata
-  {-# DEPRECATED "Use 'import Cardano.Ledger.Shelley.TxAuxData' instead" #-} (
-  module X,
-)
-where
-
-import Cardano.Ledger.Shelley.TxAuxData as X